refactor(#194):  파일이름 pagination type에서 model로 변경
