+++ apps/frontend/src/feature/pagination/index.ts
@@ -1,3 +1,3 @@
 export * from './usePagination';
 export * from './Pagination';
-export * from './type';
+export * from './model';
